requirements.txt
# Web / API ---
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.22

# UI ---
streamlit==1.41.1

# Core AI / RAG ---
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.1

# Embeddings ---
sentence-transformers==3.3.1
torch==2.2.2

# Vector Store ---
faiss-cpu==1.9.0.post1

# PDF parsing ---
pypdf==5.1.0

# Utilities ---
python-dotenv==1.0.1
pydantic==2.10.4
numpy==1.26.4

# Dev ---
pytest==8.3.4
ruff==0.8.4
